% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{calc_sup_wald_fast}
\alias{calc_sup_wald_fast}
\title{Compute the Sup-Wald statistic}
\usage{
calc_sup_wald_fast(Y, proj_info, n, hac_lag = 2L)
}
\description{
Compute the Sup-Wald statistic
}
\keyword{internal}
