% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{prepare_candidate_projections}
\alias{prepare_candidate_projections}
\title{Prepare candidate regime projections}
\usage{
prepare_candidate_projections(X, candidate_D_list, trim = 0.05)
}
\description{
Prepare candidate regime projections
}
\keyword{internal}
