% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{print.supWald.RFF}
\alias{print.supWald.RFF}
\title{Print a Sup-Wald test}
\usage{
\method{print}{supWald.RFF}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{"supWald.RFF"}.}

\item{digits}{Number of significant digits to print.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print a Sup-Wald test
}
